fix scripts for sending events to grafana
