+++ charts/federated-learning/fedbed/templates/server.yaml
@@ -22,6 +22,7 @@ spec:
     containers:
       - name: main
         image: icsforth/fedbed:latest
+        imagePullPolicy: IfNotPresent
         env:
           - name: FL_STRATEGY
             value: FedAvg
